refactor: Abstract map screen code (#655)
